list block refactoring
