+++ frontend/section04/src/app/boards/_models/detail-types.d.ts
