+++ frontend/section04/src/app/boards/_models/list-types.ts
@@ -1,0 +1,19 @@
+interface IListBoard {
+  _id: string;
+  title: string;
+  writer: string;
+  contents: string;
+  createdAt: string;
+  updatedAt: string;
+  deletedAt: string | null;
+  likeCount: number | null;
+  dislikeCount: number | null;
+  images: string[];
+  youtubeUrl: string | null;
+  __typename: string;
+}
+
+interface IBlockProps {
+  el: IListBoard;
+  idx: number;
+}
