--- frontend/section04/src/app/boards/_models/new-inputValue.d.ts
@@ -1,7 +1,0 @@
-interface IInputValue {
-  author: string;
-  password: string;
-  title: string;
-  content: string;
-  [key: string]: string;
-}
