--- frontend/section04/src/app/boards/_models/new-objects.d.ts
@@ -1,16 +1,0 @@
-interface IPlaceholderObj {
-  author: string;
-  password: string;
-  title: string;
-  content: string;
-  addressNum: string;
-  addressInput: string;
-  addressDetail: string;
-  youtube: string;
-  photo: string;
-}
-
-interface IButtonObj {
-  cancel: string;
-  register: string;
-}
